+++ yarely/qt5/__init__.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 # -*- coding: utf-8 -*-
 #
 # Copyright 2011-2019 Lancaster University.
@@ -9,11 +7,3 @@
 #
 # Licensed under the Apache License, Version 2.0.
 # For full licensing information see /LICENSE.
-
-
-import common
-
-module_name_template = "yarely.core.scheduling.manager"
-requirements = []
-
-common.launch_application(module_name_template, requirements)