+++ yarely/qt5/common/__init__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright 2011-2016 Lancaster University.
+#
+#
+# This file is part of Yarely.
+#
+# Licensed under the Apache License, Version 2.0.
+# For full licensing information see /LICENSE.
+
+
+# Local (Yarely) imports
+from yarely.qt5.common import window, renderer
+
+__all__ = ["window", "renderer"]